Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy-utils==0.41.1
bcrypt==4.0.1
pillow==10.1.0
requests==2.31.0
pyjwt==2.7.0
alembic==1.11.1
cryptography==41.0.2
celery==5.3.1
redis==4.6.0
httpx==0.24.1
opencv-python==4.8.0.74
numpy==1.25.1
pytest==7.4.0
gunicorn==21.2.0 